Refactor x/lockfile into dbnode/server
